(RQS/) Major refactoring
