+++ src/rqs/sfml_front.cpp
@@ -1,0 +1,5 @@
+//
+// Created by quothbonney on 10/30/22.
+//
+
+#include "sfml_front.h"
