+++ src/rqs/sfml_front.h
@@ -1,0 +1,14 @@
+//
+// Created by quothbonney on 10/30/22.
+//
+
+#ifndef OPENGCAS_SFML_FRONT_H
+#define OPENGCAS_SFML_FRONT_H
+
+
+class sfml_front {
+
+};
+
+
+#endif //OPENGCAS_SFML_FRONT_H
