+++ tests/CMakeLists.txt
@@ -5,8 +5,8 @@ SET(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${openGCAS_SOURCE_DIR}/bin)
 add_executable(tests
         ../tests/main.cpp
         ../tests/rqs_test.cpp
-        ../src/rqs/RasterQuery.cpp
-        ../src/rqs/rqsDataBlock.cpp
+        ../src/rqs/raster_query.cpp
+        ../src/rqs/rqs_datablock.cpp
         )
 target_link_libraries(tests PRIVATE Catch2::Catch2 GDAL::GDAL)
